{{ cookiecutter.project_name }}/src/requirements.txt
cmsplugin-filer==1.0.1
Django==1.8.13
django-appconf==1.0.1
django-classy-tags==0.7.2
django-cms==3.2.5
django-compressor==2.0
django-extensions==1.6.1
django-filer==1.1.1
django-formtools==1.0
django-mptt==0.8.3
django-polymorphic==0.8.1
django-reversion==1.8.7
django-rosetta==0.7.11
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.1.1
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-link==1.7.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.3
djangocms-video==1.0.0
easy-thumbnails==2.3
html5lib==0.9999999
microsofttranslator==0.7
mysql==0.0.1
MySQL-python==1.2.5
opbeat==3.3
Pillow==3.2.0
polib==1.0.7
pytz==2016.4
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
Unidecode==0.4.19
